+++ instalooter/cli/__init__.py
@@ -26,12 +26,12 @@ import verboselogs
 from .. import __version__
 from ..looters import InstaLooter, HashtagLooter, ProfileLooter, PostLooter
 from ..pbar import TqdmProgressBar
-from ..batch import BatchRunner
+from ..batch import BatchRunner, logger as batch_logger
 
 from . import logutils
 from .constants import HELP, USAGE, WARNING_ACTIONS
 from .time import get_times_from_cli
-from .login import login
+from .login import login, logger as login_logger
 
 
 __all__ = ["main", "logger"]
@@ -39,7 +39,6 @@ __all__ = ["main", "logger"]
 
 #: A `~logging.Logger` instance used within the `.cli` module.
 logger = verboselogs.VerboseLogger(__name__)
-logger.addHandler(logging.StreamHandler())
 
 
 @logutils.wrap_warnings(logger)
@@ -72,12 +71,13 @@ def main(argv=None, stream=None):
         _print(USAGE)
         return 0
 
-    # Set the logger up with the requested logging level
+    # Set the loggers up with the requested logging level
     level = "ERROR" if args['--quiet'] else args.get("--loglevel", "INFO")
-    coloredlogs.install(
-        level=int(level) if level.isdigit() else level,
-        stream=stream,
-        logger=logger)
+    for logger_ in (logger, login_logger, batch_logger):
+        coloredlogs.install(
+            level=int(level) if level.isdigit() else level,
+            stream=stream,
+            logger=logger_)
 
     # Check the requested logging level
     if args['-W'] not in WARNING_ACTIONS:
@@ -91,8 +91,10 @@ def main(argv=None, stream=None):
         try:
             # Run in batch mode
             if args['batch']:
+                # Load the batch configuration from the given file
                 with open(args['<batch_file>']) as batch_file:
                     batch_runner = BatchRunner(batch_file, args)
+                # Run the batch
                 batch_runner.run_all()
                 return 0
 
@@ -145,13 +147,13 @@ def main(argv=None, stream=None):
             )
 
             # Attempt to login and extract the timeframe
+            if args['--username']:
+                login(args)
+            if args['--num-to-dl']:
+                args['--num-to-dl'] = int(args['--num-to-dl'])
             try:
-                if args['--username']:
-                    login(args)
-                if args['--time']:
+                if args['--time'] is not None:
                     args['--time'] = get_times_from_cli(args['--time'])
-                if args['--num-to-dl']:
-                    args['--num-to-dl'] = int(args['--num-to-dl'])
             except ValueError as ve:
                 _print("invalid format for --time parameter:", args["--time"])
                 _print("    (format is [D]:[D] where D is an ISO 8601 date)")